Updating measure protocol for new design
