+++ dbt_semantic_interfaces/protocols/measure.py
@@ -50,25 +50,3 @@ class Measure(Protocol):
     def reference(self) -> MeasureReference:
         """Returns a reference to this measure."""
         ...
-
-
-class _MeasureMixin:
-    """Some useful default implementation details of MeasureProtocol."""
-
-    name: str
-    expr: Optional[str] = None
-    non_additive_dimension: Optional[NonAdditiveDimensionParameters] = None
-    agg_time_dimension: Optional[str] = None
-
-    @property
-    def checked_agg_time_dimension(self) -> TimeDimensionReference:
-        assert self.agg_time_dimension, (
-            f"Aggregation time dimension for measure {self.name} is not set! This should either be set directly on "
-            f"the measure specification in the model, or else defaulted to the primary time dimension in the data "
-            f"source containing the measure."
-        )
-        return TimeDimensionReference(element_name=self.agg_time_dimension)
-
-    @property
-    def reference(self) -> MeasureReference:  # noqa: D
-        return MeasureReference(element_name=self.name)
