+++ tests/test_implements_satisfy_protocols.py
@@ -5,11 +5,30 @@ from dbt_semantic_interfaces.objects.elements.dimension import (
     DimensionTypeParams,
     DimensionValidityParams,
 )
+from dbt_semantic_interfaces.objects.elements.measure import Measure
 from dbt_semantic_interfaces.objects.metadata import FileSlice, Metadata
 from dbt_semantic_interfaces.protocols.dimension import Dimension as DimensionProtocol
+from dbt_semantic_interfaces.protocols.measure import Measure as MeasureProtocol
 from dbt_semantic_interfaces.protocols.metadata import Metadata as MetadataProtocol
+from dbt_semantic_interfaces.type_enums.aggregation_type import AggregationType
 from dbt_semantic_interfaces.type_enums.dimension_type import DimensionType
 from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
+
+
+@runtime_checkable
+class RuntimeCheckableMeasure(MeasureProtocol, Protocol):
+    """We don't want runtime_checkable versions of protocols in the package, but we want them for tests."""
+
+    pass
+
+
+def test_measure_protocol() -> None:  # noqa: D
+    test_measure = Measure(
+        name="test_measure",
+        agg=AggregationType.SUM,
+        agg_time_dimension="some_time_dimension",
+    )
+    assert isinstance(test_measure, RuntimeCheckableMeasure)
 
 
 @runtime_checkable
